Add dependency on DynamicExpresso for interpreter. Remove GenerateParserRegex in favour of using GenerateMultiParamParserRegex for all. Unit tests working with DynamicExpresso but needs more testing to ensure all edge cases are working as expected.

+++ src/ReportViewer.NET/Parsers/Text/InStrParser.cs
@@ -59,7 +59,7 @@ namespace ReportViewer.NET.Parsers.Text
                 null
             );
 
-            if (parsedExpression is int)
+            if (parsedExpression is long)
             {
                 // We've been provided with start index.
                 var startIndx = parsedExpression.ExpressionAsInt();
